requirements.txt
google-api-python-client
python-dotenv
sqlalchemy
pyyaml
jupyter
google-cloud-bigquery
apache-airflow
google-cloud-bigquery==3.22.0
pandas==2.2.3
requests==2.32.3
